Refactor Recording Request and Response creation into inits methods
