+++ PeakNetwork/Core/RecordingLogger.swift
@@ -26,77 +26,17 @@ public class RecordingLogger: Logger {
         guard
             let urlRequest = idToURLRequestMap[id],
             let method = urlRequest.httpMethod,
-            let url = urlRequest.url,
-            let requestHeaders = urlRequest.allHTTPHeaderFields
+            let url = urlRequest.url
             else { return }
         
-        let requestBodyString = urlRequest.httpBody.flatMap { String(data: $0, encoding: String.Encoding.utf8) }
-        
-        let path = url.path
-        
-        let request = Recording.Request(headers: requestHeaders, body: requestBodyString)
-        let response = createResponse(from: urlResponse, data: data)
+        let request = Recording.Request(urlRequest: urlRequest)
+        let response = Recording.Response(urlResponse: urlResponse, data: data)
         let times = Recording.Times(start: requestDate, end: responseDate)
-        let recording = Recording(method: method, host: url.host, path: path, query: url.query, times: times, request: request, response: response)
+        let recording = Recording(method: method, host: url.host, path: url.path, query: url.query, times: times, request: request, response: response)
         
         writer.write(recording, toFileNamed: "\(Int(requestDate.timeIntervalSince1970)).json" )
         
     }
-    
-    private func createResponse(from response: URLResponse?, data: Data?) -> Recording.Response? {
-        
-        guard let urlResponse = response as? HTTPURLResponse else {
-            return nil
-        }
-        
-        let responseBodyString = data.flatMap { String(data: $0, encoding: String.Encoding.utf8) }
-        
-        // Refactor into init on Recording.Headers
-        let responseHeaders = urlResponse.allHeaderFields
-            .compactMap { (key, value) -> Recording.Headers? in
-                guard let stringKey = key as? String, let stringValue = value as? String else {
-                    print("Cannot convert response headers to strings")
-                    return nil
-                }
-                return [stringKey: stringValue]
-                
-            }
-            .reduce(Recording.Headers()) { (current, next) -> Recording.Headers in
-                return current.merging(next) { (first, _) in first }
-        }
-        
-        return Recording.Response(status: urlResponse.statusCode, headers: responseHeaders, body: responseBodyString)
-    }
-}
-
-public struct Recording: Codable {
-    public struct Times: Codable {
-        let start: Date
-        let end: Date
-    }
-    
-    public typealias Headers = [String: String]
-    
-    public struct Request: Codable {
-        let headers: Headers
-        let body: String?
-    }
-    
-    public struct Response: Codable {
-        let status: Int
-        let headers: Headers
-        let body: String?
-    }
-    
-    let method: String
-    let host: String?
-    let path: String
-    let query: String?
-    let times: Times
-    
-    let request: Request
-    let response: Response?
-    
 }
 
 public protocol WriteRecording {
